refactor & rework frontend to match new backend
There has been a change in how the backend returns stuff. This is now
reflected upon in the frontend.
