+++ packages/clients/textLocator/src/locales.ts
@@ -15,8 +15,6 @@ const locales: LanguageOption[] = [
     type: 'de',
     resources: {
       textLocator: {
-        // TODO temporary key, should be removed when no longer needed
-        notImplemented: 'Diese Funktion ist noch nicht implementiert.',
         layers: {
           [openStreetMap]: 'OpenStreetMap',
           [openSeaMap]: 'OpenSeaMap',
