+++ packages/clients/textLocator/src/plugins/GeometrySearch/utils/vectorDisplay.ts
@@ -5,7 +5,7 @@ import { Vector } from 'ol/source'
 import { createEmpty, extend } from 'ol/extent'
 import { geoJson, idPrefixes } from '../../../utils/coastalGazetteer/common'
 import { FeatureType, TreeViewItem } from '../types'
-import { TitleLocationFrequency } from '../../../utils/literatureByToponym'
+import { TitleLocationFrequency } from '../../../utils/textLocatorBackend/literatureByToponym'
 import { heatStyles, typeToStyle } from './vectorStyles'
 
 const vectorSource = new Vector()
