+++ packages/clients/textLocator/src/utils/textLocatorBackend/urlSuffix.ts
@@ -1,0 +1,6 @@
+export default {
+  findDocumentsByTitle: '/findDocumentsByTitle',
+  // NOTE unused atm; POST {"id": documentId } => TitleLocationFrequency
+  locationsToText: '/locationsToText',
+  lookupLocation: '/lookup/location',
+}
