--- packages/clients/textLocator/src/utils/urlSuffix.ts
@@ -1,5 +1,0 @@
-export default {
-  findDocumentsByTitle: '/findDocumentsByTitle',
-  locationsToText: '/locationsToText',
-  lookUpLocationsIndividually: '/lookup/locations_individually',
-}
